refactor(station) : Add ModelMapper and change ServiceInterface to ServiceUseCase
The StationController has dependency about StationUseCase. The StationService implements StationUseCase to dependency inversion.

The ModelMapper is to split the dependency between Entity and DTO
