+++ src/main/java/atdd/station/StationController.java
@@ -1,14 +1,9 @@
 package atdd.station;
 
-import atdd.station.entity.StationEntity;
-import atdd.station.service.IStationService;
-import atdd.station.service.StationDTO;
-import atdd.station.service.StationService;
-import java.net.URI;
-import javax.servlet.http.HttpServlet;
+import atdd.station.usecase.StationUseCase;
+import atdd.station.usecase.StationDTO;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,7 +14,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 @RestController
 public class StationController {
     @Autowired
-    IStationService stationService;
+    StationUseCase stationService;
 
     @PostMapping("/stations")
     public ResponseEntity<StationDTO> addStation(@RequestBody @Valid StationDTO stationDTO){
