--- src/main/java/atdd/station/service/IStationService.java
@@ -1,6 +1,0 @@
-package atdd.station.service;
-
-public interface IStationService {
-  StationDTO addStation(StationDTO stationDTO);
-
-}
