+++ src/main/java/atdd/station/usecase/StationUseCase.java
@@ -1,0 +1,5 @@
+package atdd.station.usecase;
+
+public interface StationUseCase {
+  StationDTO addStation(StationDTO stationDTO);
+}
